requirements.txt
asgiref==3.8.1
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.5
defusedxml==0.7.1
dj-database-url==1.0.0
Django==5.2.3
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
djoser==2.3.3
gunicorn==23.0.0
idna==3.10
numpy==2.2.6
oauthlib==3.3.1
packaging==25.0
pandas==2.3.1
pillow==11.2.1
psycopg2==2.9.10
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python3-openid==3.2.0
pytz==2025.2
requests==2.32.4
requests-oauthlib==2.0.0
six==1.17.0
social-auth-app-django==5.5.1
social-auth-core==4.7.0
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
whitenoise==6.9.0
